Added valve crud operations

+++ src/main/java/com/example/watering_system/data/repository/ValveRepository.java
@@ -3,4 +3,4 @@ package com.example.watering_system.data.repository;
 import com.example.watering_system.data.entity.Valve;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ValveRepository extends JpaRepository<Valve, Long> {}
+public interface ValveRepository extends JpaRepository<Valve, Integer> {}
